requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic-settings==2.7.0
python-dotenv==1.0.1
anthropic==0.40.0
msal==1.31.1
slack-bolt==1.21.3
sqlalchemy==2.0.36
aiofiles==24.1.0
httpx==0.28.1
PyJWT==2.10.1
